implement fit_polyridge; rename source file to match
